# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: xLemma Protocol API
  version: 0.2.0
  description: >-
    Reference API for decentralized researcher identities, proof-production jobs,
    independent Lean verification, PoIR observations, compute quotes, x402
    payment-protected resources, novelty review, and lemma capsules.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/verification-jobs:
    post:
      operationId: createVerificationJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVerificationJob"
      responses:
        "201":
          description: Job created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VerificationJob" }
  /v1/verification-jobs/{jobId}:
    get:
      operationId: getVerificationJob
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Current job state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VerificationJob" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/verification-jobs/{jobId}/observations:
    post:
      operationId: submitObservation
      parameters:
        - $ref: "#/components/parameters/JobId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: ../schemas/observation.schema.json }
      responses:
        "200": { description: Observation appended }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/verification-jobs/{jobId}/payment-required:
    post:
      operationId: createVerificationPaymentOffer
      summary: Construct an x402 V2 payment-required offer for a verification job
      parameters:
        - $ref: "#/components/parameters/JobId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PaymentOfferRequest" }
      responses:
        "402": { $ref: "#/components/responses/PaymentRequired" }
        "404": { description: Verification job not found }
  /v1/verification-jobs/{jobId}/evaluate:
    post:
      operationId: evaluateVerificationJob
      parameters:
        - $ref: "#/components/parameters/JobId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [policy]
              properties:
                policy: { $ref: ../schemas/formal-consensus-policy.schema.json }
      responses:
        "200": { description: Formal consensus outcome }
...
components:
  parameters:
...
    JobId:
      in: path
      name: jobId
      required: true
      schema: { type: string }
  schemas:
...
    CreateVerificationJob:
      type: object
      required: [researcher_id, claim_id, theory_id, artifact_id, policy_id, maximum_budget_minor_units, settlement_asset]
      properties:
        researcher_id: { type: string }
        claim_id: { type: string }
        theory_id: { type: string }
        artifact_id: { type: string }
        policy_id: { type: string }
        maximum_budget_minor_units: { type: integer, minimum: 1 }
        settlement_asset: { type: string }
    VerificationJob:
      type: object
      additionalProperties: true
      required: [job_id, state, created_at, updated_at]
      properties:
        job_id: { type: string }
        state:
          type: string
          enum: [draft, claim_committed, quoted, funded, assigned, formalizing, candidate_ready, building, checkers_committed, checkers_revealed, passed, failed, divergent, challenged, finalized, rejected, quarantined, published, revalidated, superseded]
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PaymentOfferRequest:
      type: object
      additionalProperties: false
      required: [compute_quote_id, amount, asset, network, pay_to, max_timeout_seconds, artifact_commitment, model_policy, rights_manifest_hash, revenue_route_hash, delivery_mode, valid_until]
      properties:
        compute_quote_id: { type: string, pattern: "^xlquote:blake3:[0-9a-f]{64}$" }
        amount: { type: string, pattern: "^[1-9][0-9]*$" }
        asset: { type: string, minLength: 1 }
        network: { type: string, minLength: 1 }
        pay_to: { type: string, minLength: 1 }
        max_timeout_seconds: { type: integer, minimum: 1 }
        artifact_commitment: { type: string, minLength: 16 }
        model_policy: { type: string, minLength: 1 }
        rights_manifest_hash: { type: string, minLength: 16 }
        revenue_route_hash: { type: string, minLength: 16 }
        delivery_mode: { type: string, minLength: 1 }
        valid_until: { type: string, format: date-time }
  responses:
    PaymentRequired:
      description: x402 payment authorization required
      headers:
        PAYMENT-REQUIRED:
          schema: { type: string }
          description: Base64-encoded x402 V2 JSON with the xlemma extension
...
    NotFound:
      description: Resource not found
    Conflict:
      description: State or identity conflict